cli-next: Refuse to build w/o a 'node_modules' dir
Summary:
The previous version of the Airtable Blocks CLI produced an error if the
end-user attempt to run the development server or release the code in
the absence of the `node_modules` directory. Update the corresponding
commands in cli-next to preserve this behavior.

The fundamental file system operation (detecting the presence of a
directory) had previously been repeated twice elsewhere in the codebase,
though those existing versions silently tolerated exceptional
circumstances (tending to obscure problems should they arise). Introduce
a utility function which abstracts the operation and apply it
consistently.

Test Plan:
1. Visit the directory of a valid App
2. Remove the `node_modules` directory
3. Run the following command: `block run`
4. Verify the CLI prints an error message which describes the absence of
   the `node_modules` directory
5. Verify that the CLI does *not* report information about the App
   development server
6. Verify the CLI exists with a non-zero exit code
7. Run the following command: `block release`
8. Verify the CLI prints an error message which describes the absence of
   the `node_modules` directory
9. Verify that the CLI does *not* report information about releasing the
   App
10. Verify the CLI exists with a non-zero exit code

Reviewers: benrazon, mzgoddard, somehats

Reviewed By: benrazon, mzgoddard

Differential Revision: https://airtable.phacility.com/D30364