cli-next: add submit command
Summary:
This diff adds the submit command. It finds the source code dependencies with
the bundler, packages those files on disk along with the files under the
project's root determined by the location of the block.json config file. Then
confirms with the developer if they are ready to submit the package. Once
confirmed it uploads the zip to airtable's servers.

The package may include dependencies not under the project's root, if it does so
it nests the files in the zip archive under the highest common path. If there
are files `/a/b/block.json`, `/a/b/c.js` and `/a/d/e.js`, the zip will contain paths
`b/block.json`, `b/c.js` and `d/e.js`. In either case it includes an
AIRTABLE_BLOCK_ROOT_PATH file that has in it the relative path to the project's root.

Test Plan:
- `block init BLOCK_IDENTIFIER cli-next-submit-test-app`
- `cd cli-next-submit-test-app`
- `block submit`
- `N` when prompted
- `block submit`
- `y` when prompted (this will upload the package to Airtable)

Reviewers: somehats, benrazon, jugglinmike

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D30469