cli-next: Validate API key & API key name
Summary:
Extend the `set-api-key` sub-command to validate the API key and (when
specified) corresponding API key name. Describe the key error using the
term "malformed" to distinguish it from the previously-introduced error
relating to rejected API keys. (That error is produced in the context of
failed API interactions, and since it directs users to the `set-api-key`
subcommand, it is inappropriate for use in this context.)

Preserve the existing behavior which repeatedly prompts the user when
reading the key via the standard input stream. This patch does *not*
include an automated test for that behavior due to a deficiency in the
testing infrastructure. A subsequent commit will address this issue and
introduce an automated test.

Test Plan:
1. Run the command `block set-api-key`
2. Verify that the process prompts for an API key
3. Enter the value: ky123
4. Verify that the process reports that the provided key is invalid
5. Verify that the process prompts for an API key again
6. Enter the value: key123
7. Verify that the process reports that the key has been updated
8. Verify that the process exits with a status code of zero
9. Verify that the API key has been correctly stored in the system's
   user configuration file
10. Run the command `block set-api-key key123 --api-key-name ???`
11. Verify that the process reports that the provided key name is
    invalid
12. Verify that the process exits with a status code of zero
13. Verify that the API key has *not* been stored in the system's user
    configuration file
15. Run the command `block set-api-key key123 --api-key-name a`
16. Verify that the process reports that the key has been updated
17. Verify that the process exits with a status code of zero
18. Verify that the API key has been correctly stored in the system's
    user configuration file

Reviewers: mzgoddard, somehats, jugglinmike, minicat, fredz

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D30833