Support accessing and watching base color
Summary:
Depends on D31271

Allow blocks to access the color of the base they're installed in, and watch for changes.

Test Plan: Once D31271 is live on prod and we do a release of the sdk, I can release a test block that should display the color of the base and responsively change as the base color changes.

Reviewers: #team-platform-ecosystem, SeanKeenan-at

Reviewed By: #team-platform-ecosystem, SeanKeenan-at

Subscribers: minicat, SeanKeenan-at, s2tephen

Differential Revision: https://airtable.phacility.com/D31273