DPTF: handle block submit
Summary:
https://airtable.com/tblSbpimM8Spp9txp/viw2RPmJhvFySGtPg/recV7uirBdkEYxTCM?blocks=hide
This uses the newly added blockV2 routes for uploading code submissions.

Test Plan:
Create a v2 block
Run block submit
Observe that the corresponding records were created in the block submission base and that
the code is in s3.
Note that testing this might not be possible in staging since we might not have created
a corresponding base as go/carb

Reviewers: minicat, fredz

Reviewed By: fredz

Subscribers: #proj-dptf

Differential Revision: https://airtable.phacility.com/D31424