Make yarn install not update package-lock.json
Summary:
I noticed when running yarn install on a clean branch we would modify the package-lock to include
blocks-testing. I believe this is because the example blocks were using an older version of the
package so we couldn't use the version within the repo.

Test Plan: Run yarn install and observe that no files are updated anymore

Reviewers: minicat, somehats

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D31604