Cli next: refactor api class
Summary:
https://airtable.com/tblSbpimM8Spp9txp/viw2RPmJhvFySGtPg/rec0rXhvvxhP3SM0r?blocks=hide
This makes the AirtableApi class abstract and moves some shared functionality into it so
that we can create an appropriate api class and use it directly instead of needing the extra
upload helper classes. It also moves the api options into the constructor so that they don't
need to be passed to every method.

Test Plan: release and submit a v1 and v2 block

Reviewers: minicat, fredz

Reviewed By: minicat, fredz

Differential Revision: https://airtable.phacility.com/D31649