SDK: expose record limits
Summary:
https://airtable.com/tblSbpimM8Spp9txp/viw2RPmJhvFySGtPg/recflYpqm3Fooi4WY?blocks=hide
This is the corresponding change on the sdk side for D32060. This is needed to that blocks
can correctly compute the maximum number of records that can be created.

Test Plan:
Create a custom app
Add console.log(base.getMaxRecordsPerTable());
Run the app and observe that it prints 50000
In the application admin flag screen in support panel set CUSTOM_MAX_NUM_ROWS_PER_TABLE to a value
greater than 50k for the application you are using to run your app in
Observe that it prints the new number

Reviewers: billylittlefield, SeanKeenan-at

Reviewed By: billylittlefield

Differential Revision: https://airtable.phacility.com/D32065