CLI: handle patches during release
Summary:
In order to get csv import to work with cli-next we needed to use patch package to update one
of its dependencies. Unfortunately, this broke our automated releases to staging because
block release uses isolated builds by default, but npm install won't succeed if it attempts to
run patch package without any patch files. This simply adds a step to also copy the patches
directory if it exists to work around this.

Test Plan: Run block release for a block that does and doesn't use patch-package and observe that it succeeds

Reviewers: minicat, fredz

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D32595