Add EXTERNAL_SYNC_SOURCE field type
Summary: Depends on D28770

Test Plan:
Tested by running base schema and vega lite with a local copy of the sdk.
Base schema shows the field type when hovering over the field and vega lite can show a bar graph of an EXTERNAL_SYNC_SOURCE field and autocompletes the field name and type.

Reviewers: somehats, minicat, raghavsethi

Reviewed By: minicat

Subscribers: minicat

Differential Revision: https://airtable.phacility.com/D29289