Cli: swap cli-package names to allow releasing cli-next to beta
Summary:
We're ready to start the cli-next beta. However, since we want to release it as a beta version
of the same package and yarn workspaces don't support multiple packages with the same name
we need to temporarily rename the current cli. This means that every time we release a new
version of one of the cli's we'll need to check the package name and potentially make a similar
diff to update the name. I'm not too worried about us forgetting to do this since the push should
fail if the name doesn't match an npm package

Test Plan:
Running yarn succeeds and I can run blocks locally with either cli. I'm not aware of any
testing we can perform on the npm publish or git mirror'ing other than trying it.

Reviewers: minicat, billylittlefield

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D32912