Base schema: package lock
Summary:
Now that we've published a new version of the blocks sdk, running npm install updates the
package-lock.json file. This shouldn't actually matter as we handle this case when creating
a block from an example, but I needed to do this to get the block to pass qa, so I figure that
I should check in the change.

Test Plan: Run the block and observe that sync source fields now display their type correctly

Reviewers: minicat, SeanKeenan-at

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D33025