support nested absolute paths
Summary:
### Summary
Preparation for converting relative => absolute imports for blocks inside of hyperbase. Previously, the CLI had an assumption that the `absoluteImportRoot` was only one layer deep. As it stands, the only block that utilizes `absoluteImportRoot` was scripting (with `absoluteImportRoot = "scripting"` - it's subdirectory in hyperbase), so the assumption was valid.

However, we now want to use `absoluteImportRoot` for blocks where the import root will be nested (e.g. `"blocks/summary"`). As it stands now, some logic will fail upon `block release`, stating that no such directory exists. This makes the logic more robust so as to support a nested `absoluteImportRoot`.

It also ensures that the legacy `'airtable-block'` module is created for non-isolated builds, which is necessary if we want to use the `--disable-isolated-builds` flag on unmigrated blocks.

This change unblocks the relative import conversions for blocks hyperbase (e.g. D36053)

### Example / context
Currently, blocks utilize relative imports:
```
import ChartBlock from './chart_block';
```

As part of the continued hyperbase style assimilation, we'd like to convert these to relative imports:
```
import ChartBlock from 'blocks/chart/frontend/chart_block';
```

This was previously made possible for scripting by a new hidden property in block.json: `__hyperbase.absoluteImportRoot: "scripting"`. This value is consumed the blocks CLI, which symlinks the output transpiled content to a newly-created folder at `<output_transpiled_stuff>/node_modules/<absolute_import_root>` - a strategy similar to how absolute imports are resolved for the broader hyperbase codebase.

This code is currently calling `mkdir` on a resolved path in a non-recursive way, so if the `absoluteImportRoot` is nested ("blocks/chart"), we'll hit an error along the lines of "You asked me to create `a/b/c`, but `a/b` doesn't exist". This previously wasn't an issue for scripting, where the root was only a single layer deep.

This code also fixes a separate (but relevant) issue where we don't write the stubbed `airtable-block` module in non-isolated builds, which prevents us from doing `--disable-isolated-build` on the 4 unmigrated blocks. By writing this module all the time, we can alleviate the issue.

Test Plan: Once the new CLI version is published, verify that scripting app can be released using v0.0.65 of the CLI

Reviewers: fredz, kevinwilde, #team-platform-ecosystem

Reviewed By: fredz, #team-platform-ecosystem

Differential Revision: https://airtable.phacility.com/D36039