[Block - URL Preview] Loom video fix
Summary:
Loom URL's have a bug because they https://loom.com/... redirects to http://www.loom.com/... which redirects to https://www.loom.com/...

In the middle of all of this you might notice a redirect to an http:// - which triggers a CORS error, as we don't allow loading non-secure domains. The fix is to skip the redirects and change the URL to include www.

npm install also caused an update to the package.lock, so I rolled that into this

{F338368}
{F338369}

Test Plan: Loom URL's should now work inside of URL Preview, previously they would not load

Reviewers: #team-platform-ecosystem, billylittlefield

Reviewed By: #team-platform-ecosystem, billylittlefield

Subscribers: billylittlefield

Differential Revision: https://airtable.phacility.com/D36176