sdk: add description to createFieldAsync
Summary:
SDK counterpart to D36895

Adds `description` to `Table.createFieldAsync`.

While it's optional on the liveapp side (for backwards compatibility with previous versions of the SDK), I decided to make the type of `description` in the mutation non-optional. I don't feel too strongly about this decision, but it seemed nice to enforce that "null" is the value for an empty description, rather that also being `undefined`. Empty string is allowed, since empty string is OK on the liveapp side too (e.g. editing description to be empty via UI writes `''`.).

On the other hand, it's optional in the public API so that this isn't a breaking change, and for usability.

Test Plan: Tested manually with a test block, and updated the unit tests

Reviewers: billylittlefield, kevinwilde

Reviewed By: kevinwilde

Differential Revision: https://airtable.phacility.com/D36904