[Blocks SDK] Cache Field configs
Summary:
When debugging a CSV import issue with Tim we discovered that computing `airtableInterface.fieldTypeProvider.getConfig` can be rather expensive in some cases. In particular when a multi-select field has a LOT of options, it has to loop through all of them and generate a large object.

For the CSV Import app this problem was compounded as it would loop over all 15,000 records in the CSV and then fetch the type on every field. Since some fields had ~100's of options, this was looping over ~millions of select field options. This change would cut it back down to hundreds!

Also field.config() fetches this twice today! Now we'll only be fetching it once due to the caching.

Test Plan:
Ensure that fields are properly updated inside of apps that observe this behavior, especially when changing a field type, or field config options.

These things are to some extent already covered by existing unit tests however.

Reviewers: timdeng2324, minicat, #team-platform-ecosystem

Reviewed By: minicat, #team-platform-ecosystem

Differential Revision: https://airtable.phacility.com/D38070