Force Yarn Install in sdk package (#47)
I'm never sure in various repos if we use npm or yarn. This explicitly makes this requirement part of the package.json. (I believe we're supposed to use yarn with package.json - but if I'm wrong and it's npm - even more reason we should have this here!)

This does add a small amount of time to installation, as it first downloads the utility which checks if the package installer.

This is what running `npm install` looks like:
![image](https://user-images.githubusercontent.com/72051418/134111903-72365bc4-3c14-455b-893f-d55c8fae822b.png)