Blocks-testing git repo is incorrect (#49)
You can see that the link on our [npm page is broken](https://www.npmjs.com/package/@airtable/blocks-testing).

I'm unsure how to specify the specific subdirectory, but this mirrors how our blocks-sdk npm page looks. We probably should update the root Readme to specify that that there are multiple packages in the repo, since currently it's a symlink to only the sdk's readme. I'll do that in a separate commit.