Shared code example improvements (#48)
This is in response to some feedback [from Ronan](https://community.airtable.com/t/blocks-cli-v2-beta-how-to-use-code-from-other-directories/43197/4).

The readme is hopefully self-explanatory!

I also plan on responding to Ronan in the thread with this: (just edited because he responded recently)
```
Hey @Ronen_Babayoff (I work with @Emma_Yeap)

First off - thank you for your feedback; it's very helpful!

We published the example as a [stand-alone git repo](https://github.com/Airtable/apps-shared-code) for easier access in the future.

I also just add `@airtable/blocks` as a peer dependency now. We know that *yarn* workspaces work - but have not tested with npm workspaces. Glad to hear that npm workspaces also work!

Completely agreed that we also should add instructions how how to set up `npm link` here - especially since it works well with hot-reloads, I did that and added you to the readme!

When you said that `npm link` is only working on `index.tsx`, do you mean it doesn't work for `index.jsx` when not working in a typescript project - or do you mean when specifying the transpiled `index.tsx` target? (Aka - does npm link just not work at all in non-ts projects).

The sub-module import problem is interesting - makes sense that it's on Typescripts side 👍.

Regarding the release issue - could you share your code with us, or a minimal repro of the issue so that we can help debug?
```

(This is as soon as the git repo is actually public from IT - hopefully tomorrow