Example with blocks api (#61)
Ronen correctly pointed out that we include blocks - but don't actually use them in the shared code. This adds a toy example here.

He also mentioned a hooks violation - it turns out that this happens because of multiple floating react installations - he must have added react as a devDependency/dependency in the shared code. This attempts to solve this problem by hoisting the install up one layer - see the readme for more.

I also updated the link to the npm page - and use the channel name, so that it will remain up-to-date with future releases