Release process: Only mirror the package being published (#128)
# Summary
Currently, releasing a new version of the SDK will also include unpublished changes to cli-next or blocks-testing in the public release commit. This is undesirable and makes it difficult to tell what has actually changed in a release. For example, if 
- we are on cli-next@beta.4
- we are actively making changes to cli-next without publishing
- we are actively making changes to the sdk and publishing those as new sdk releases

then, 
- the unpublished changes to cli-next are now being included in the public sdk release commits
- when we finally release cli-next@beta.5, the diff will not show all the changes since beta.4 since a bunch of them were already on master from the sdk release commits

Example: https://github.com/Airtable/blocks/commit/f386ffd0604a6ee14829497668fb2e69bb34336b

I've moved the git-mirror.json files into each package so that they can include only the package directory that should be synced when it's published. The global files are included for all, so publishing any of the 3 packages may change the global files. For now, the git-mirror.json files are just copy-pasted, with only the package directory changed. It would be easy to add support for "extending" a base config like you can in a tsconfig if we wanted to in the future.

## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [X] Not needed
  - [ ] Added in this PR



# Test Plan

I added a dryRun argument to the git-mirror commands and ran locally with this flag to verify that the correct files were included.


Reviewed By: SeanKeenan-at, JamesMoody-at
GitHub Pull Request: https://github.com/Hyperbase/hyperbase/pull/128