Remove invariant checks for commentCount and createdTime (#132)
# Summary
There is an escalation [here](https://airtable.com/appWh5G6JXbHDKC2b/tblJvlqwVOgRLmgdT/viwmeRnBauWem7uJY/recIwhESIa1LIhW82?blocks=hide) where the developer of the Data Fetcher app encounters an error (very rarely) causes an invariant to trigger.

This diff removes the invariant checks for two reasons:
 - These checks are already gone on the experimental branch (they do not apply in a world where record state is not stored inside the SDK)
 - This check only catches a rare edge case - which I believe to be transient, and does not catch all transient inconsistencies of this form. See below for more.

Some notes about how this can occur / what I've tried:

 - Created at time can change legally, optimistic updates within the SDK specify the created_at time, which can (and does) change after the record has been committed
 - This code path that checks the invariant is only exercised when loading specific fields that have not yet been loaded
I suspect that this is a race condition where:
   - Record is created (optimistically)
   - Record is updated in hyperbase, but not yet sent across bridge
   - Code subscribes to some new set of fields (which returns the new created_at time)
   - Record created_at time is realtime pushed across the bridge
 - Why is this happening? Not sure. Potential resolutions:
     - Fix the source of the race condition (ideal)
     - Remove the invariants (why is it important to invariant this? If we can prove it should be eventually consistent anyways the worst case is that some newly loaded fields are "newer" than other already loaded fields - but these invariants don't _prevent_ this, they just crash in some subset of cases) - what this solution is
     - Upgrade to experimental SDK (experimental SDK is not ready for something as heavy use as Data Fetcher)
     - Fetch all of the records earlier, or don't fetch a specific slice of fields. (does not set oneself up for experimental sdk / is a hack that would need to be implemented inside data fetcher)

In full transparency I did attempt to create a block that reproduces the above failure case but was not successful. I was also not able to reproduce the issue in the shared datafetcher base Andy provided.

Of note in the above case the "inconsistent data" that can happen is:
 - Fields A, B are fetched
 - Cell in Field A, B, and C are changed
 - Fields B, C are fetched
 - At this point B and C are up to date, A is out of date
 - Field A and B are delivered to SDK through realtime mechanism change
 - Everything is consistent
This inconsistency case _is not_ caught by the above assertions, it only catches this in the special case where B is a commentCount or a createdTime - both of which are fields which do not change often.


## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [x] Not needed
  - [ ] Added in this PR

# Test Plan

No additional testing is required.

Reviewed By: minicat
GitHub Pull Request: https://github.com/Hyperbase/hyperbase/pull/132