eslint-plugin-blocks: Copy over implementation getStaticPropertyName instead of using internal eslint file (#139)
# Summary
Slack thread: https://h.slack.com/archives/C018YBVKELD/p1644598968151529
`@airtable/eslint-plugin-blocks` does not play nice with the latest version of eslint because we were relying on an internal file in the package. This copies over the implementation of the function we needed rather than importing it.

The plan is to publish a new version of `@airtable/eslint-plugin-blocks` and then update usages in hyperbase to unblock the hyperbase eslint upgrade.

## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [x] Not needed
  - [ ] Added in this PR



# Test Plan
Automated tests


Reviewed By: JamesMoody-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/139