Blocks SDK: Check if LinkedRecordQueryResult is deleted before the invariant check for isValid (#146)
# Summary

While QA testing performance improvements to the Page designer block, two crashes were uncovered that both have to do with linked records:

1. App crash when record containing linked records is deleted and then restored
https://share.support.airtable.com/JruOP7KO

2. App crashes when linked record field is deleted
https://share.support.airtable.com/P8u1ZRZZ

After some digging, it looks like the root cause of these crashes are actually stemming from a bug in the SDK that was introduced in version 1.2.1 where we (Issue 1) are not reseting `isValid` to true while restoring deleted records that contain linked record fields and (Issue 2) are somehow setting `isValid` to false before we finish deleting linked record fields. 

There's a bigger issue to still be addressed here that has existed since 2020 but not many blocks are at 1.2.1 or higher to run into it and of those, I believe page designer is the only block that could potentially run into this. While this code change does prevent the crash and ultimately allows us to follow the intended path (exit when `!this.isDataLoaded`), given the comments around `isValid`, it seems like we explicitly expect for this attribute to be false for deleted fields by design so it seems like a more exact fix is needed here.

After talking with a few folks on the team, I'm inclined to try to move forward with this workaround to 1. stop the (potential) bleeding for any apps that this may already be affecting or could in the future following an SDK bump and 2. to help finally get page designer unblocked for publishing. I've tried seeing if downgrading to 1.2.0 in the meantime would work, but there is a fix in 1.7.1 that is also required to prevent a bug from manifesting in page designer (or any other app it might affect). Additionally, I tried using the experimental SDK but the issue appears to exists there as well.

With all that said, I welcome any other ideas, opposing opinions, ~scolding~, or potential better fixes anyone might have in mind for this. Otherwise, I'll get this released as a new version and update page designer to use it while continuing to try to find a better fix.

## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [ ] Not needed
  - [x] Added in this PR

# Test Plan

This will be tested as part of general QA of the change to Page designer [here](https://h.slack.com/archives/CCNFR41JP/p1646087746619159), generally we'll need to confirm the following:

- Apps (specifically Page designer) should not crash when a record containing a linked record field is deleted and then restored
- Apps (specifically Page designer) should not crashes when linked record field is deleted


Reviewed By: SeanKeenan-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/146