Cli-next: properly handle errors in dangerouslyCrossSpawnAndReturnTrimmedOutputAsync (#154)
# Summary

I was testing on my windows machine and discovered an edge case we weren't properly handling. If git is not installed on the developer's machine, the release command was crashing. (So not windows-specific, but only discovered because I was testing on a machine other than my work laptop!)

The reason we execute a git command during `block release` is so that our source maps work properly. See [this diff](https://airtable.phacility.com/D25285). I'm not sure if this is still relevant as we move to Sentry instead of Rollbar; I didn't look into whether we also use the git hash when initializing Sentry.

This is only relevant for 1st party blocks, so having git installed should not be a requirement for a blocks developer. Our existing code wrapped the entire body of `getGitHash` in a try/catch (presumably for this reason), but this didn't actually handle errors when trying to get the git hash. The issue is that when getting the git hash, we spawn a child process which emits errors "out of band" in a sense. You can't handle them with a try/catch, you need to use `child.on('error', errHandler)`. This PR does makes that change.

## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [X] Not needed
  - [ ] Added in this PR



# Test Plan

After this change, I'm able to `block release` from a machine without git.


Reviewed By: FredZhao-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/154