cli-next: smoke tests (#147)
# Summary
This change adds the smoke tests from the legacy cli to cli-next. It includes a bit of a refactor of the copied test to make use of typescript. We also needed to update a few of the copied .bat files for running tests on windows to match the usage for cli-next.

## Tasks
- Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md)
  - [x] Not needed
  - [ ] Added in this PR

# Test Plan

- Pre-release smoke test should run successfully on mac
- Pre-release smoke test should run successfully on pc
- Post-release smoke test should run successfully on mac
- Post-release smoke test should run successfully on pc


Reviewed By: kevinwilde-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/147