sdk: rename apps to extensions in examples (#179)
# Summary
As part of the 'Apps' to 'Extensions' rename, this change updates the example docs to use the correct noun. Since we don't have access to the getNounForBlock() helper function from the blocks-sdk repo, we'll need to land the updated docs in the hyperbase repo after the feature flag has been enabled.

## Tasks

-   Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md) or
    [the CLI changelog](packages/cli-next/CHANGELOG.md)
    -   [ ] Not needed
    -   [x] Added in this PR

# Test Plan




Reviewed By: SeanKeenan-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/179