chore(cliNext): manually update cli-next CHANGELOG.md and fix copy in README.md (#219)
# Summary

For cli-next we need to manually update the CHANGELOG.md file. Following the release of 2.0.4, this includes the new version and updates in the change log.

Additionally, this make a small update to the copy of README.md to change one more reference of "API key"

## Tasks

-   Add to the changelog if needed? I.e. [the SDK changelog](packages/sdk/CHANGELOG.md) or
    [the CLI changelog](packages/cli-next/CHANGELOG.md)
    -   [ ] Not needed
    -   [x] Added in this PR

# Test Plan




Reviewed By: HongdaJiang-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/219