upgrade typescript and other deps for vuln management (#241)
# Summary
 
**Why**: 
The main reason for this PR is to remove all the critical dependency vulnerabilities as discovered by `yarn audit`. 

Originally, the intention was to do a narrower upgrade of only the `release-it` deps in the root `packages.json` file, but CI would fail because various type errors would occur across the packages (e.g. `/packages/sdk`).  These type errors are pre-existing, and seemingly have to have existed for a long time: they are reproducible even before upgrading `release-it`;  the impact is that we cannot merge in any fixes touching the root `package.json` because it'll fail CI. Therefore, decided to upgraded the typescript version (and various other dependencies) along the way. 

`yarn audit` before this PR:
![image](https://github.com/Hyperbase/blocks-sdk/assets/726490/69102354-b08e-4dd2-997d-d35a12400807)


`yarn audit` after the changes in this PR:
![image](https://github.com/Hyperbase/blocks-sdk/assets/726490/50e1836a-dcf9-4b3f-bfc8-c69bb5e11dd5)

A couple notes: 
* The immediate goal is to just publish the legacy CLI tool to npm for now, as that is our committed goal for the quarter (e.g. `packages/cli`). The SDK has not been tested yet, and we should do so before publishing to npm . After this PR merges, we will publish the legacy CLI to npm. 
* You can see the various type errors that were occurring in this slack thread: https://h.slack.com/archives/C018YBVKELD/p1719332716895639 

**What**:
* Upgraded to typescript v5.4.5 across the various `packages/**` in this monorepo. 
* A few other targeted deps upgrades (e.g. `eslint`, etc)
* Attempt to fix some type errors. For some of the trickier ones, just decided to do `@ts-ignore` for now to make progress here. We should attempt to fix these later. 



# Test Plan

* Manually test legacy CLI (e.g. v0.0.N which maps to `packages/cli`). After this PR merges, we will publish the legacy CLI to npm. 
* TODO: All other `packages/**` need to still be tested before publishing to npm. 

Reviewed By: ClementHoang-at
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/241