Vuln management of examples/ blocks (#247)
# Summary

Resolve security vulnerabilities in the `examples/` blocks. These blocks
are open source blocks we publish to the community as examples for
extensions development. Some of these blocks we also make available in
airtable.com/marketplace (e.g. vega lite extension).

A couple notable things about this PR:
* Completely deletes the `ui-playground` block because I didn't notice
it available as an open source block in github.com/Airtable , nor do I
see it in the marketplace. It was difficult to resolve all the security
dependencies for it (and it was also a flow-typed block), so I just
removed it for now.
* In `todo-block` and `update-records-block` removed the
`test/jest.setup.js` files which was providing shim workaround out
because I think it's unnecessary for jest v29.7.0 ? (it removes the deps
to `@sheerun/mutationobserver-shim`)


# Test Plan

some manual testing for some of the blocks, but didn't do an exhaustive
test.