CLI: Set default region for backend route requests in development (#272)
When running blocks in development, we should set a reasonable default region for "local" requests to backend blocks. Currently that region is set to `''`, this change updates it to `''` in order to match the default blocks region defined by development config / credentials in hyperbase.

Reviewed By: richsinn
GitHub Pull Request: https://github.com/Hyperbase/blocks-sdk/pull/272