CLI: Ignore more files from npm (#278)
The legacy CLI package uses an `.npmignore` file which causes the
default behavior of ignoring files from `.gitignore` to be skipped.
Because of this, the published npm package includes `node_modules/` and
several other directories which are unnecessary in the published npm
package. The inclusion of `node_modules` created problems when using
pnpm to install the CLI globally, resulting in the `Cannot find module
'har-validator'` error message.