Interface Extensions: custom properties (#293)
# Summary
This PR adds support for custom properties for interface blocks to the
SDK.

I'm not 100% set on the exact API. In this PR, the signature is
```ts
useCustomProperties(params: {customProperties: Array<BlockPageElementCustomProperty>}): {
    customPropertyValueByKey: {[key: string]: unknown};
    errorState: {error: Error} | null
}
```

I'd like to land this PR with that signature and then we can always opt
to change it in a subsequent PR before the interface blocks SDK is
stable.

See corresponding hyperbase changes at: 
- https://github.com/Hyperbase/hyperbase/pull/176847
- https://github.com/Hyperbase/hyperbase/pull/176963

# Test Plan
Manually tested with corresponding hyperbase changes.