Flow doesn't work anymore in CI (#298)
# Summary
All builds in CI are failing due to flow not working on the machines
used for CI. Flow is only used for the old CLI and an internal changelog
tool. I'm just going to stop running flow in those packages because it's
blocking us and we're not even working in those packages.

Failing main build: https://buildkite.com/airtable/blocks-sdk/builds/427

Error is 
```
Error: Platform not supported.
   at Object.<anonymous> (/workdir/tools/changelog-publish/node_modules/flow-bin/cli.js:17:9)
```
That line is

https://github.com/flow/flow-bin/blob/e527bf1624a16b00e7d3202671c0c2c38578ca57/cli.js#L11
which is requiring this file:

https://github.com/flow/flow-bin/blob/e527bf1624a16b00e7d3202671c0c2c38578ca57/index.js#L12-L19

I tried upgrading flow-bin because the latest version of the index.js
file has support for `process.platform === 'linux' && process.arch ===
'arm64'` but that led to
```
/workdir/packages/cli/node_modules/flow-bin/flow-linux-arm64-v0.268.0/flow: /lib/aarch64-linux-gnu/libm.so.6: version `GLIBC_2.38' not found (required by /workdir/packages/cli/node_modules/flow-bin/flow-linux-arm64-v0.268.0/flow)
```

## Tasks

- Add to the changelog if needed? I.e. [the SDK
changelog](packages/sdk/CHANGELOG.md) or
    [the CLI changelog](packages/cli-next/CHANGELOG.md)
    -   [x] Not needed
    -   [ ] Added in this PR

# Test Plan
CI